first draft of collection rebuild
